+++ DemoSpiritsAPI/Controllers/SpiritController.cs
@@ -4,6 +4,7 @@ using SpiritsClassLibrary.DTOs.SpiritDTOs;
 using SpiritsClassLibrary.Models;
 using DemoSpiritsAPI.Servicies.Interfaces;
 using Microsoft.AspNetCore.Mvc;
+using DemoSpiritsAPI.Servicies;
 
 namespace DemoSpiritsAPI.Controllers
 {
@@ -13,10 +14,13 @@ namespace DemoSpiritsAPI.Controllers
     {
 
         private ISpiritService _spiritService;
+        private IGoogleAuthService _googleAuthService;
 
-        public SpiritController(ISpiritService spiritService)
+
+        public SpiritController(ISpiritService spiritService, IGoogleAuthService googleAuthService)
         {
             _spiritService = spiritService;
+            _googleAuthService = googleAuthService;
         }
         [HttpGet(Name = "GetAllSpirits")]
         public IActionResult GetAll()
@@ -34,6 +38,11 @@ namespace DemoSpiritsAPI.Controllers
         [HttpPatch(Name = "SetupTestData")]
         public IActionResult SetupTestData()
         {
+            var authResult = _googleAuthService.ValidateAdminPermission(Request);
+            if (authResult.Exception != null)
+            {
+                return BadRequest(authResult.Exception.Message);
+            }
             var result = _spiritService.SetupTestData();
             if (result.Exception == null)
             {
@@ -59,7 +68,11 @@ namespace DemoSpiritsAPI.Controllers
         [HttpDelete("{id}", Name = "DeleteSpirit")]
         public IActionResult Delete(int id)
         {
-
+            var authResult = _googleAuthService.ValidateAdminPermission(Request);
+            if (authResult.Exception != null && !authResult.Result)
+            {
+                return BadRequest(authResult.Exception.Message);
+            }
             var result = _spiritService.Delete(id);
 
             if(result.Exception == null)
@@ -73,8 +86,19 @@ namespace DemoSpiritsAPI.Controllers
 
         [HttpPost(Name = "CreateSpirit")]
         public IActionResult Create([FromBody] CreateSpiritDTO createSpiritDTO)
-        { 
-          
+        {
+            var authResult = _googleAuthService.ValidateAdminPermission(Request);
+            if (authResult.Exception != null && !authResult.Result)
+            {
+                Exception ex = authResult.Exception;
+                string message = "";
+                while(ex != null)
+                {
+                    message += ex.Message + "\n";
+                    ex = ex.InnerException;
+                }
+                return BadRequest(/*authResult.Exception.Message*/message);
+            }
             var result =  _spiritService.Create(createSpiritDTO);
             if(result.Exception == null)
             {
@@ -87,7 +111,11 @@ namespace DemoSpiritsAPI.Controllers
         [HttpPut(Name = "UpdateSpirit")]
         public IActionResult Update([FromBody] UpdateSpiritDTO updateSpiritDTO)
         {
-
+            var authResult = _googleAuthService.ValidateAdminPermission(Request);
+            if (authResult.Exception != null && !authResult.Result)
+            {
+                return BadRequest(authResult.Exception.Message);
+            }
             var result = _spiritService.Update(updateSpiritDTO);
             if(result.Exception == null)
             {
